tools: move some utility functions to lib/utils.py
